tilp: replace GLADE_CB macro with TILP_EXPORT.
The GLADE_CB macro was already custom as it needed to be modified for our use of -fvisibility=hidden.
It was defined in an otherwise libglade-specific header, so the new tilp-specific macro was placed in tilp_defs.h, so that the glade specific files can be removed in a later commit.

Signed-off-by: Jon Sturm <jonimoose@gmail.com>

git-svn-id: http://svn.tilp.info/repos/tilp@4451